Сделала кривую авторизацию через контекст
